requirements.txt
absl-py==2.3.1
astunparse==1.6.3
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-auth==2.45.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
joblib==1.5.3
keras==2.14.0
kiwisolver==1.4.9
libclang==18.1.1
lz4==4.4.5
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
ml-dtypes==0.2.0
mtcnn==1.0.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
# tensorboard==2.14.1
# tensorboard-data-server==0.7.2
# tensorflow==2.14.1 # CUDA version  11.8
# tensorflow-estimator==2.14.0
# tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
wrapt==1.14.2
